+++ src/interfaces/GenericEvent.ts
