+++ src/interfaces/Job.ts
