+++ src/interfaces/NoteEvent.ts
