+++ src/interfaces/NoteEventAttributes.ts
