+++ src/interfaces/Pipeline.ts
