+++ src/interfaces/Plugin.ts
