+++ src/interfaces/Storage.ts
